feat: configure the number of players shown on the board. defaults to 10
Teams webhook messages have a hard limit of 28 KB. This limit is easily reached
with as few as 15 players. This change introduces a safe default of 10, and allows
the user to configure with a new environment variable. If the board has more
players than the limit, a link to the web board will be displayed.

BREAKING_CHANGE: Leaderboards in Teams will show only the top 10 players

Due to Team's webhook message size limits, boards sent to Teams will now show
a maximum of 10 players. This limit can be configured by setting
`LEADERBOARD_LIMIT`, but values higher than 15 will likely result in messages
failing to reach Teams.

Fixes #6